后台辅助工具接口提交2

+++ backstage/src/main/java/com/line/backstage/entity/OrderInfo.java
@@ -147,9 +147,9 @@ public class OrderInfo extends QueryRequest implements Serializable {
     @ApiModelProperty("数据状态")
     private Integer del;
     private Integer diyId = 0;
-    private Integer inPoint;
-    private Integer outPoint;
-    private Integer integral;
+    private Double inPoint;
+    private Double outPoint;
+    private Double integral;
     private Double subMoney;
     private Double resultMoney;
     private String skuName;
